+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `requiredFields` property to the form set `*` in label of required fields
+
 ## [0.10.0] - 2023-04-27
 
 ### Added